Move location + path helpers to separate file
